pyproject.toml
[tool.poetry]
name = "djproj"
version = "0.1.0"
description = ""
authors = ["omid9h <omid9h@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
attrs = "^22.2.0"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
django-environ = "^0.10.0"
django-filter = "^22.1"
Inject = "^4.3.1"
Pillow = "^9.4.0"
whitenoise = "^6.3.0"
celery = [
    { version = "^5.2.7" },
    { version = "^5.2.7", extras = [
        "redis",
    ], markers = "extra == 'redis'" },
]
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
flower = "^1.2.0"

[tool.poetry.extras]
redis = ["celery"]

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
autopep8 = "^2.0.1"
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"
autoflake = "^2.0.1"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# override black's default line-length
line-length = 99
exclude = [".git", ".venv"]

[tool.isort]
# make it compatible with black
profile = "black"
